+++ core/src/main/java/com/backblaze/b2/client/B2StorageClient.java
@@ -37,6 +37,7 @@ import com.backblaze.b2.client.structures.B2ListFileVersionsRequest;
 import com.backblaze.b2.client.structures.B2ListKeysRequest;
 import com.backblaze.b2.client.structures.B2ListPartsRequest;
 import com.backblaze.b2.client.structures.B2ListUnfinishedLargeFilesRequest;
+import com.backblaze.b2.client.structures.B2Part;
 import com.backblaze.b2.client.structures.B2StartLargeFileRequest;
 import com.backblaze.b2.client.structures.B2StoreLargeFileRequest;
 import com.backblaze.b2.client.structures.B2UpdateBucketRequest;
@@ -247,9 +248,9 @@ public interface B2StorageClient extends Closeable {
      *
      * This method assumes you have already called startLargeFile(). The return value
      * of that call needs to be passed into this method. However, this method will
-     * currently call finish file.
+     * call finish file.
      *
-     * XXX: should we switch to letting the caller finish the large file?
+     * XXX: add storePartsForLargeFileFromLocalContent that stores parts but doesn't finish the file.
      *
      * @param fileVersion The B2FileVersion for the large file getting stored.
      *                    This is the return value of startLargeFile().
@@ -276,9 +277,9 @@ public interface B2StorageClient extends Closeable {
      *
      * This method assumes you have already called startLargeFile(). The return value
      * of that call needs to be passed into this method as part of a
-     * B2StoreLargeFileRequest. However, this method will currently call finish file.
+     * B2StoreLargeFileRequest. However, this method will call finish file.
      *
-     * XXX: should we switch to letting the caller finish the large file?
+     * XXX: add storePartsForLargeFileFromLocalContent that stores parts but doesn't finish the file.
      *
      * @param storeLargeFileRequest The B2StoreLargeFileRequest for the large file
      *                              getting stored. This is built from the return
@@ -371,22 +372,20 @@ public interface B2StorageClient extends Closeable {
     /**
      * Stores a large file, where storing each part may involve different behavior
      * or byte sources.
-     *
+     * <p>
      * For example, this method supports the use case of making a copy of a file
      * that mostly has not changed, and the user only wishes to upload the parts
      * that have changed. In this case partStorers would be a mix of
      * B2CopyingPartStorers and one or more B2UploadingPartStorers.
-     *
+     * <p>
      * Another use case would be reattempting an upload of a large file where some
      * parts have completed, and some haven't. In this case, partStorers would
      * be a mix of B2AlreadyStoredPartStorer and B2UploadingPartStorers.
-     *
+     * <p>
      * This method assumes you have already called startLargeFile(). The return value
      * of that call needs to be passed into this method. However, this method will
-     * currently call finish file. Note that each part, whether copied or uploaded,
-     * is still subject to the minimum part size.
-     *
-     * XXX: should we switch to letting the caller finish the large file?
+     * call finish file. Note that each part, whether copied or uploaded, is still
+     * subject to the minimum part size.
      *
      * @param fileVersion The B2FileVersion for the large file getting stored.
      *                    This is the return value of startLargeFile().
@@ -411,19 +410,19 @@ public interface B2StorageClient extends Closeable {
      * Stores a large file, where storing each part may involve different behavior
      * or byte sources, optionally allowing caller to pass SSE-C parameters to match
      * those given to startLargeFile().
-     *
+     * <p>
      * For example, this method supports the use case of making a copy of a file
      * that mostly has not changed, and the user only wishes to upload the parts
      * that have changed. In this case partStorers would be a mix of
      * B2CopyingPartStorers and one or more B2UploadingPartStorers.
-     *
+     * <p>
      * Another use case would be reattempting an upload of a large file where some
      * parts have completed, and some haven't. In this case, partStorers would
      * be a mix of B2AlreadyStoredPartStorer and B2UploadingPartStorers.
-     *
+     * <p>
      * This method assumes you have already called startLargeFile(). The return value
      * of that call needs to be passed into this method as part of a
-     * B2StoreLargeFileRequest. However, this method will currently call finish file.
+     * B2StoreLargeFileRequest. However, this method will call finish file.
      * Note that each part, whether copied or uploaded,
      * is still subject to the minimum part size.
      *
@@ -443,6 +442,84 @@ public interface B2StorageClient extends Closeable {
      * @throws B2Exception If there's trouble.
      */
     B2FileVersion storeLargeFile(
+            B2StoreLargeFileRequest storeLargeFileRequest,
+            List<B2PartStorer> partStorers,
+            B2UploadListener uploadListenerOrNull,
+            ExecutorService executor) throws B2Exception;
+
+    /**
+     * Stores large file parts, where storing each part may involve different behavior
+     * or byte sources.
+     * <p>
+     * For example, this method supports the use case of preparing to make a copy
+     * of a file that mostly has not changed, and the user only wishes to upload
+     * the parts that have changed. In this case partStorers would be a mix of
+     * B2CopyingPartStorers and one or more B2UploadingPartStorers.
+     * <p>
+     * Another use case would be reattempting some part uploads for an unfinished
+     * large file for which some parts have completed, and some haven't. In this case,
+     * partStorers would be a mix of B2AlreadyStoredPartStorer and B2UploadingPartStorers.
+     * <p>
+     * This method assumes you have already called startLargeFile(). The return value
+     * of that call needs to be passed into this method. Note that each part, whether
+     * copied or uploaded, is still subject to the minimum part size. Note also that
+     * this method does <b>not</b> finish the large file.
+     *
+     * @param fileVersion The B2FileVersion for the large file for which these parts
+     *                    are being uploaded. This is the return value of startLargeFile().
+     * @param partStorers The list of objects that know how to store the part
+     *                    they are responsible for.
+     * @param uploadListenerOrNull The object that handles upload progress events.
+     *                             This may be null if you do not need to be notified
+     *                             of progress events.
+     * @param executor The executor for uploading parts in parallel. The caller
+     *                 retains ownership of the executor and is responsible for
+     *                 shutting it down.
+     * @return The list of parts after all have been stored.
+     * @throws B2Exception If there's trouble.
+     */
+    List<B2Part> storePartsForLargeFile(
+            B2FileVersion fileVersion,
+            List<B2PartStorer> partStorers,
+            B2UploadListener uploadListenerOrNull,
+            ExecutorService executor) throws B2Exception;
+
+    /**
+     * Stores large file parts, where storing each part may involve different behavior
+     * or byte sources, optionally allowing caller to pass SSE-C parameters to match
+     * those given to startLargeFile().
+     * <p>
+     * For example, this method supports the use case of making a copy of a file
+     * that mostly has not changed, and the user only wishes to upload the parts
+     * that have changed. In this case partStorers would be a mix of
+     * B2CopyingPartStorers and one or more B2UploadingPartStorers.
+     * <p>
+     * Another use case would be reattempting some part uploads for an unfinished
+     * large file for which some parts have completed, and some haven't. In this case,
+     * partStorers would be a mix of B2AlreadyStoredPartStorer and B2UploadingPartStorers.
+     * <p>
+     * This method assumes you have already called startLargeFile(). The return value
+     * of that call needs to be passed into this method as part of a
+     * B2StoreLargeFileRequest. Note that each part, whether copied or uploaded,
+     * is still subject to the minimum part size. Note also that this method does
+     * <b>not</b> finish the large file.
+     *
+     * @param storeLargeFileRequest The B2StoreLargeFileRequest for the large file
+     *                              for which these parts are being stored. This is
+     *                              built from the return value of startLargeFile() and
+     *                              any other relevant parameters.
+     * @param partStorers The list of objects that know how to store the part
+     *                    they are responsible for.
+     * @param uploadListenerOrNull The object that handles upload progress events.
+     *                             This may be null if you do not need to be notified
+     *                             of progress events.
+     * @param executor The executor for uploading parts in parallel. The caller
+     *                 retains ownership of the executor and is responsible for
+     *                 shutting it down.
+     * @return The list of parts after all have been stored.
+     * @throws B2Exception If there's trouble.
+     */
+    List<B2Part> storePartsForLargeFile(
             B2StoreLargeFileRequest storeLargeFileRequest,
             List<B2PartStorer> partStorers,
             B2UploadListener uploadListenerOrNull,